refactor to packages fma.mmar and fma.rs
