+++ .gitignore
@@ -5,3 +5,4 @@
 .vscode
 .DS_Store
 **.pyc
+fma.egg-info/
